+++ session/tests/fixtures/fresh-boilerplate/routes/_app.tsx
@@ -1,0 +1,11 @@
+import { define } from "../utils.ts";
+
+export default define.page(function App({ Component, state, url }) {
+  return (
+    <html>
+      <body>
+        <Component />
+      </body>
+    </html>
+  );
+});